Fix circular includes
